refactor (typing indicator): move into a reusable element
this will allow message replies to also trigger the typing indicator, without code duplication
