+++ src/main.rs
@@ -39,6 +39,7 @@ use crate::components::main;
 use crate::components::prelude::{auth, loading, unlock};
 
 pub mod components;
+pub mod iui_kit;
 pub mod iutils;
 pub mod language;
 pub mod themes;
